move and copy utility functions
